Elements added to products

+++ database/migrations/2015_10_05_121358_create_stocks_table.php
@@ -21,7 +21,7 @@ class CreateStocksTable extends Migration
                 ->on('products');
 
             $table->integer('amount');
-            $table->double('cost');
+            $table->decimal('cost');
 
             $table->timestamps();
         });
